app/backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
# AI dependencies
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.5.2

# Utilities
joblib==1.3.2

catboost==1.2.5
